requirements.txt
flask
flask_sqlalchemy
flask_jwt_extended
flask_bcrypt
pytest
pandas
openpyxl
pytest-cov